Work around for missing Sermon Descriptions
Sermon Manager previously stored the sermon/post description into a meta field 'sermon_description'. The migration process to newer versions is very unreliable and buggy. As such here is a work around that will query the meta field and display it if content is available there instead of the_content();